requirements.txt
asgiref
attrs
certifi
charset-normalizer
coreapi
coreschema
dj-database-url
Django
django-cors-headers
django-environ
django-phonenumber-field
djangorestframework
djangorestframework-simplejwt
drf-yasg
gunicorn
idna
inflection
itypes
Jinja2
jsonschema
MarkupSafe
packaging
phonenumbers
Pillow
pip
psycopg2
PyJWT
pyotp
pyparsing
pyrsistent
pytz
PyYAML
requests
ruamel.yaml
ruamel.yaml.clib
setuptools
six
sqlparse
swagger-spec-validator
twilio
uritemplate
urllib3
whitenoise
xlwt